move PaypalError object creation in the model
